created four separte modules for the cart.
